+++ src/handlers/list-handler.ts
@@ -1,0 +1,5 @@
+import { InitCommandOptions } from "../common/typings";
+
+export const run = (opts: InitCommandOptions) => {
+  throw new Error("Not implemented");
+};
